@@ -27,18 +27,29 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  http: ^0.12.0+4
+  cupertino_icons: ^1.0.2
+  webview_flutter: ^2.0.10
+  flutter_search_bar: ^3.0.0-dev.1
+  font_awesome_flutter: ^9.1.0
+  flutter_launcher_icons: ^0.9.1
+  material_design_icons_flutter: ^4.0.5955
+  http: ^0.13.3
   cached_network_image: ^2.0.0
   carousel_slider: ^4.0.0
   curved_navigation_bar: ^1.0.1 #latest version
-  animated_theme_switcher: ^2.0.1 
+  animated_theme_switcher: ^2.0.1
   flutter:
     sdk: flutter
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  webview_flutter: ^2.0.10
+  ionicons: ^0.1.2
+  community_material_icon: ^5.9.55
+  web_scraper: ^0.1.4
+  firebase_auth: ^3.0.1
+  cloud_firestore: ^2.4.0
+  firebase_core: ^1.4.0
+  provider: ^5.0.0
+  flutter_spinkit: ^5.0.0
+  convex_bottom_bar: ^3.0.0
+  material_floating_search_bar: ^0.3.4
 
 dev_dependencies:
   flutter_test:
