@@ -1,8 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'components/ad_service.dart';
+import 'components/ad2.dart';
 import 'components/banner.dart';
-import 'components/news_view.dart';
 
 class MyHome extends StatefulWidget {
   const MyHome({Key? key}) : super(key: key);
@@ -15,8 +14,7 @@ class _MyHomeState extends State<MyHome> {
   @override
   Widget build(BuildContext context) {
     return ListView(
-      shrinkWrap: true,
-      physics: const ClampingScrollPhysics(),
+      physics: ClampingScrollPhysics(),
       children: <Widget>[
         const banner(),
         const SizedBox(
@@ -25,25 +23,14 @@ class _MyHomeState extends State<MyHome> {
         Container(
           margin: const EdgeInsets.only(left: 20),
           child: const Text(
-            'Our Services',
+            'Connection - Services',
             style: TextStyle(
               fontSize: 25,
               fontWeight: FontWeight.w500,
             ),
           ),
         ),
-        const ad_service(),
-        Container(
-          margin: const EdgeInsets.only(left: 20),
-          child: const Text(
-            'Health News',
-            style: TextStyle(
-              fontSize: 25,
-              fontWeight: FontWeight.w500,
-            ),
-          ),
-        ),
-        const news_view(),
+        const ad_service()
       ],
     );
   }
