@@ -1,14 +1,13 @@
-import 'package:curved_navigation_bar/curved_navigation_bar.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:medo/order/order.dart';
-import 'package:medo/personal/personal.dart';
-import 'package:medo/search/search.dart';
-import 'package:medo/serviceInfomation/service_information.dart';
+import 'package:medo/models/the_user.dart';
+import 'package:medo/screens/wrapper.dart';
+import 'package:medo/services/auth.dart';
+import 'package:provider/provider.dart';
 
-import 'component/app_bar.dart';
-import 'home/home.dart';
-
-void main() {
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
   runApp(const myApp());
 }
 
@@ -17,67 +16,15 @@ class myApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      title: 'Medo - Medicine & Doctor',
-      debugShowCheckedModeBanner: false,
-      home: myAppPage(),
+    return StreamProvider<TheUser?>.value(
+      catchError: (_, err) => null,
+      initialData: null,
+      value: AuthService().user,
+      child: const MaterialApp(
+        title: 'Medo - Medicine & Doctor',
+        debugShowCheckedModeBanner: false,
+        home: Wrapper(),
+      ),
     );
-  }
-}
-
-class myAppPage extends StatefulWidget {
-  const myAppPage({Key? key}) : super(key: key);
-  @override
-  _myAppPageState createState() => _myAppPageState();
-}
-
-class _myAppPageState extends State<myAppPage> {
-  final MyHome _myHome = const MyHome();
-  final MyOrder _myOrder = const MyOrder();
-  final MySearch _mySearch = const MySearch();
-  final MyService _myService = const MyService();
-  final MyPersonal _myPersonal = const MyPersonal();
-
-  // Widget _showPage =  myApp();
-  Widget _showPage = const MyHome();
-
-  Widget _pageChooser(int page) {
-    switch (page) {
-      case 0:
-        return _myHome;
-      case 1:
-        return _mySearch;
-      case 2:
-        return _myOrder;
-      case 3:
-        return _myService;
-      case 4:
-        return _myPersonal;
-      default:
-        return const Center(child: Text('Can not find the page'));
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: const CustomAppBar(),
-        body: Container(child: _showPage),
-        backgroundColor: Colors.white,
-        bottomNavigationBar: CurvedNavigationBar(
-          items: const <Widget>[
-            Icon(Icons.home, size: 30),
-            Icon(Icons.search_rounded, size: 30),
-            Icon(Icons.add, size: 30),
-            Icon(Icons.medical_services, size: 30),
-            Icon(Icons.supervised_user_circle_outlined, size: 30),
-          ],
-          backgroundColor: Colors.white,
-          onTap: (index) {
-            setState(() {
-              _showPage = _pageChooser(index);
-            });
-          },
-        ));
   }
 }
