@@ -1,20 +1,20 @@
 import 'dart:convert';
-// ignore: import_of_legacy_library_into_null_safe
 import 'package:http/http.dart' as http;
+import 'package:medo/component/key.dart';
 import 'article_model.dart';
 
 class News {
   List<Article> news = [];
 
   Future<void> getNews() async {
-    String url =
-        "https://newsapi.org/v2/top-headlines?country=us&category=health&apiKey=277a2ecb12a34db1a1bfa13057a72329";
+    String myLink =
+        "https://newsapi.org/v2/top-headlines?country=us&category=health&apiKey=${api_key}";
 
-    var response = await http.get(url);
+    final response = await http.get(Uri.parse(myLink));
 
     var jsonData = jsonDecode(response.body);
 
-    if (jsonData['status'] == "ok") {
+    if (jsonData['status'] == 'ok') {
       jsonData["articles"].forEach((element) {
         if (element['urlToImage'] != null && element['description'] != null) {
           Article article = Article(
