@@ -2,8 +2,12 @@ import 'package:animated_theme_switcher/animated_theme_switcher.dart';
 import 'package:flutter/material.dart';
 import 'package:medo/personal/components/profile_list_item.dart';
 import 'package:medo/personal/components/header.dart';
+import 'package:medo/screens/homeScreen/infoscreen.dart';
+import 'package:medo/services/auth.dart';
 
 class ProfileScreen extends StatelessWidget {
+  final AuthService _auth = AuthService();
+
   @override
   Widget build(BuildContext context) {
     return ThemeSwitchingArea(child: Builder(builder: (context) {
@@ -15,10 +19,16 @@ class ProfileScreen extends StatelessWidget {
             SizedBox(height: 50),
             Header(),
             SizedBox(height: 20),
-            ProfileListItem(
-              icon: Icons.account_circle_rounded,
-              text: 'Personal Information',
-              hasNavigation: true,
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.blue,
+              ),
+              child: Text('Personal Information',
+                  style: TextStyle(color: Colors.white)),
+              onPressed: () {
+                Navigator.push(context,
+                    MaterialPageRoute(builder: (context) => InfoScreen()));
+              },
             ),
             ProfileListItem(
               icon: Icons.favorite_outline_rounded,
@@ -35,10 +45,14 @@ class ProfileScreen extends StatelessWidget {
               text: 'Help & Support',
               hasNavigation: true,
             ),
-            ProfileListItem(
-              icon: Icons.logout_rounded,
-              text: 'Logout',
-              hasNavigation: false,
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                primary: Colors.blue,
+              ),
+              child: Text('Log out', style: TextStyle(color: Colors.white)),
+              onPressed: () async {
+                await _auth.signOut();
+              },
             ),
           ]))
         ],
