@@ -1,7 +1,11 @@
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_search_bar/flutter_search_bar.dart';
+import 'package:material_floating_search_bar/material_floating_search_bar.dart';
+import 'package:medo/search/components/searchOK.dart';
 
 class MySearch extends StatefulWidget {
-  const MySearch({ Key? key }) : super(key: key);
+  const MySearch({Key? key}) : super(key: key);
 
   @override
   _MySearchState createState() => _MySearchState();
@@ -10,8 +14,53 @@ class MySearch extends StatefulWidget {
 class _MySearchState extends State<MySearch> {
   @override
   Widget build(BuildContext context) {
-    return const Text('search');
+    return FloatingSearchBar(
+      hint: 'Search for Medicine, Doctor,...',
+      scrollPadding: const EdgeInsets.only(top: 16, bottom: 56),
+      transitionDuration: const Duration(milliseconds: 800),
+      transitionCurve: Curves.easeInOut,
+      physics: const BouncingScrollPhysics(),
+      openAxisAlignment: 0.0,
+      debounceDelay: const Duration(milliseconds: 500),
+      onQueryChanged: (query) {
+        // Call your model, bloc, controller here.
+      },
+      onSubmitted: (query) {
+        Navigator.push(
+            context,
+            MaterialPageRoute(
+                builder: (context) => SearchComplete(nameOfMedicine: query)));
+      },
+      // Specify a custom transition to be used for
+      // animating between opened and closed stated.
+      transition: CircularFloatingSearchBarTransition(),
+      actions: [
+        FloatingSearchBarAction(
+          showIfOpened: false,
+          child: CircularButton(
+            icon: const Icon(Icons.search),
+            onPressed: () {},
+          ),
+        ),
+        FloatingSearchBarAction.searchToClear(
+          showIfClosed: false,
+        ),
+      ],
+      builder: (context, transition) {
+        return ClipRRect(
+          borderRadius: BorderRadius.circular(8),
+          child: Material(
+            color: Colors.white,
+            elevation: 4.0,
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: Colors.accents.map((color) {
+                return Container(height: 112, color: color);
+              }).toList(),
+            ),
+          ),
+        );
+      },
+    );
   }
 }
-
-//da lam xong
